Move the Logging type into local to work around Python codegen issue 16221
